[BH-819] Fix illegal dependencies in bsp
Break circular dependencies in the BSP module. Fix some missing
dependencies which surfaced after fixing BSP.
