+++ module-bsp/board/linux/CMakeLists.txt
@@ -47,6 +47,8 @@ target_include_directories(
                 >
 )
 
+target_link_libraries(module-bsp PRIVATE sys-common)
+
 set_source_files_properties(board.cpp PROPERTIES COMPILE_FLAGS -Wno-unused-function)
 
 target_compile_options(module-bsp PUBLIC "-Wno-unused-result")
