+++ module-bsp/board/rt1051/bluetooth/BluetoothCommon.cpp
@@ -1,7 +1,6 @@
 // Copyright (c) 2017-2021, Mudita Sp. z.o.o. All rights reserved.
 // For licensing, see https://github.com/mudita/MuditaOS/LICENSE.md
 
-#include "BluetoothWorker.hpp"
 #include "bsp/bluetooth/Bluetooth.hpp"
 #include <log/log.hpp>
 #include "FreeRTOS.h"
