+++ module-bsp/hal/include/hal/battery_charger/AbstractBatteryCharger.hpp
@@ -5,7 +5,6 @@
 
 #include <FreeRTOS.h>
 #include <queue.h>
-#include <Service/Service.hpp>
 
 #include <memory>
